requirements.txt
streamlit==1.39.0
pandas==2.2.3
numpy==1.26.4
prophet==1.3.0
statsmodels==0.14.4
scikit-learn==1.5.2
xgboost==2.1.2
imbalanced-learn==0.12.4
shap==0.46.0
sentence-transformers==3.2.1
chromadb==0.5.20
langchain-text-splitters==0.3.2
groq==0.13.0
plotly==5.24.1
matplotlib==3.9.2
seaborn==0.13.2
beautifulsoup4==4.12.3
trafilatura==1.12.2
requests==2.32.3
jupyter==1.1.1
faiss-cpu==1.9.0
